Correct test_update_person() and update_person() per API docs
**The Cisco Spark API docs state:**
“Include all details for the person. This action expects all user
details to be present in the request. A common approach is to first GET
the person's details, make changes, then PUT both the changed and
unchanged values.”

Update the test case to reflect this recommendation.

Given the simple, stateless and atomic wrapping of the API calls that
this package (ciscosparkapi) provides; don’t modify the package code to
provide this functionality.  Such stateful automations can be made in
the more feature-rich ciscosparksdk package.