Refactor Response Codes, Exceptions & Automatic Rate-Limit Handling
Move the response code “data” as distributed throughout the package to
a dedicated response-code module.

Refactor the package exceptions to enable greater reuse through
inheritance and create a new SparkRateLimitError to be used when
catching rate-limit errors.

Refactor the automatic rate limit handling to be controlled with a
simple wait_on_rate_limit on/off setting, and to use the new
SparkRateLimitError for catching rate limit messages.