Add property and methods to expose JSON data
Ensure the SparkData object's JSON data is stored as an `OrderedDict` (and not a `dict`) to preserve the attribute ordering from the Spark responses.

Add a `json_data` property and `to_dict()` and `to_json()` methods to the SparkData base class to enable access to the data object's JSON data.

Note:  The json_data should return a copy of the object's data to prevent the object data from being mutated.