Update Generator Container interfaces, variables and docstrings
Update the Generator Container interfaces and docstrings to improve clarity.  The parameters and internal variable names should clearly illustrate what data types are accepted as parameters, contained, and returned.

These containers are taking on a an extra degree of complexity in the ciscosparksdk package where the containers are being extended to provide caching capabilities.  Clarity of operations is key to understanding the proper operation of these containers and their caching enabled sub-classes.