Set minimum retry-after wait time to 1 second
We believe Cisco Spark is returning a `Retry-After` of `0` most probably due to a millisecond rounding issue,  which would mean that the `Retry-After` time should really fall somewhere between `0 < t < 1`.

To account for the expectation that some amount of wait time was expected before the request should be retried, change any `Retry-After` responses of `0` seconds to `1` second to provide a minimum wait time for requests to be retried.