Update WebexTeamsDateTime
Attempting to apply a timezone to a naive datetime (no tzinfo) raises a Value error in Python v2.7 and v3.5.

ValueError: astimezone() cannot be applied to a naive datetime

Check to see if the datetime has time zone information (tzinfo). If it has timezone info, use the `astimezone()` method to adjust it to ZuluTime.  If it does not have timezone info, assume the time is UTC time and add the ZuluTimeZone (raise a warning when this happens).