Bug fix regarding pagination
https://github.com/CiscoDevNet/webexteamssdk/issues/168

The Webex Teams api does not return all parameters when returning the next url link header. Consequently,
the object listing can be broken or worse, can succeed without returning objects expected by the provided filter.

Example: a bot needs to specify mentionedPeople=me to be allowed to list room messages. If there are several pages to be returned
then the listing will fail because the mentionedPeople query parameter won't be specified when fetching the next pages

Signed-off-by: xXraphXx <xXraphXx@users.noreply.github.com>